Read border points from file
+ Split new parameterizers in two executables
