+++ flatmap/code/surf_cgal/CMakeLists.txt
@@ -27,7 +27,8 @@ if ( NOT Boost_FOUND )
 
 endif()
 
-create_single_source_cgal_program( "Flatten_Authalic_custom.cxx" )
+create_single_source_cgal_program( "Flatten_Authalic_convex.cxx" )
+create_single_source_cgal_program( "Flatten_Authalic_parametric.cxx" )
 create_single_source_cgal_program( "Flatten_Authalic.cxx" )
 create_single_source_cgal_program( "Flatten_Authalic_Iterative.cxx" )
 create_single_source_cgal_program( "Nearest_KNN.cxx" )
@@ -41,7 +42,8 @@ find_package(Eigen3 3.1.0) #(requires 3.1.0 or greater)
 include(CGAL_Eigen3_support)
 if(TARGET CGAL::Eigen3_support)
 
-    target_link_libraries(Flatten_Authalic_custom PUBLIC CGAL::Eigen3_support)
+    target_link_libraries(Flatten_Authalic_convex PUBLIC CGAL::Eigen3_support)
+    target_link_libraries(Flatten_Authalic_parametric PUBLIC CGAL::Eigen3_support)
     target_link_libraries(Flatten_Authalic PUBLIC CGAL::Eigen3_support)
     target_link_libraries(Flatten_Authalic_Iterative PUBLIC CGAL::Eigen3_support)
     target_link_libraries(Nearest_KNN PUBLIC CGAL::Eigen3_support)
@@ -60,7 +62,8 @@ endif()
 # OpenMP for Eigen
 find_package(OpenMP)
 if(OpenMP_CXX_FOUND)
-    target_link_libraries(Flatten_Authalic_custom PUBLIC OpenMP::OpenMP_CXX)
+    target_link_libraries(Flatten_Authalic_convex PUBLIC OpenMP::OpenMP_CXX)
+    target_link_libraries(Flatten_Authalic_parametric PUBLIC OpenMP::OpenMP_CXX)
     target_link_libraries(Flatten_Authalic PUBLIC OpenMP::OpenMP_CXX)
     target_link_libraries(Flatten_Authalic_Iterative PUBLIC OpenMP::OpenMP_CXX)
 endif()
